requirements.txt
google-generativeai
requests
python-dotenv
feedparser
beautifulsoup4
